add a message to describe why a key is blocked
